pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lora-router"
version = "0.1.0"
description = "Intelligent LoRA adapter routing, composition, and serving"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "Krishnam Gupta", email = "krishnam.gupta94@gmail.com" }]
keywords = ["lora", "peft", "routing", "adapter", "llm", "mixture-of-experts"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "transformers>=4.36",
    "peft>=0.8",
    "sentence-transformers>=2.2",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "faiss-cpu>=1.7",
]

[project.optional-dependencies]
serve = ["fastapi>=0.104", "uvicorn>=0.24", "httpx>=0.25"]
eval = ["matplotlib>=3.8", "rouge-score>=0.1", "tabulate>=0.9", "nltk>=3.8"]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
train = ["unsloth>=2024.1", "datasets>=2.16", "trl>=0.7"]
all = ["lora-router[serve,eval,dev]"]

[project.urls]
Homepage = "https://github.com/krishnam94/lora-router"
Repository = "https://github.com/krishnam94/lora-router"
Issues = "https://github.com/krishnam94/lora-router/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/lora_router"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501", "N806", "N803", "B905", "TC001", "TC003"]

[tool.ruff.lint.isort]
known-first-party = ["lora_router"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = false
init_typed = false
warn_required_dynamic_aliases = false

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (no GPU, no network)",
    "integration: Integration tests (may need network)",
    "gpu: Tests requiring GPU",
]
addopts = "-m 'not integration and not gpu' --tb=short"

[tool.coverage.run]
source = ["lora_router"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
